+++ src/confluent_sql/connection.py
@@ -36,9 +36,10 @@ def connect(  # noqa: PLR0913
     environment: str,
     compute_pool_id: str,
     organization_id: str,
-    cloud_provider: str,
-    cloud_region: str,
+    cloud_provider: str | None = None,
+    cloud_region: str | None = None,
     database: str | None = None,
+    host: str | None = None,
     dbname: str | None = None,  # deprecated, use database parameter
     result_page_fetch_pause_millis: int = 100,
     http_user_agent: str | None = None,
@@ -52,10 +53,15 @@ def connect(  # noqa: PLR0913
         environment: Environment ID
         compute_pool_id: Compute pool ID for SQL execution
         organization_id: Organization ID
-        cloud_provider: Cloud provider (e.g., "aws", "gcp", "azure")
-        cloud_region: Cloud region (e.g., "us-east-2", "us-west-2")
+        cloud_provider: Cloud provider (e.g., "aws", "gcp", "azure"). Ignored if host parameter is
+            provided.
+        cloud_region: Cloud region (e.g., "us-east-2", "us-west-2"). Ignored if host parameter is
+            provided.
         database: The default Flink database (Kafka cluster) to use when resolving
             table/view/udf names (optional)
+        host: The base URL for Confluent Cloud API (optional). If not provided, the host will be
+            constructed based on the cloud_provider and cloud_region parameters in the format
+            "https://flink.{cloud_region}.{cloud_provider}.confluent.cloud".
         dbname: Deprecated alias for database parameter (optional)
         result_page_fetch_pause_millis: Maximum milliseconds to wait between fetching pages of
             statement results (per statement). Defaults to 100ms. Prevents tight loops of requests
@@ -88,11 +94,12 @@ def connect(  # noqa: PLR0913
     if not organization_id:
         raise InterfaceError("Organization ID is required")
 
-    if not cloud_provider:
-        raise InterfaceError("Cloud provider is required")
+    if not host:
+        if not cloud_provider:
+            raise InterfaceError("Cloud provider is required when host is not provided")
 
-    if not cloud_region:
-        raise InterfaceError("Cloud region is required")
+        if not cloud_region:
+            raise InterfaceError("Cloud region is required when host is not provided")
 
     if not flink_api_key or not flink_api_secret:
         raise InterfaceError("Flink API key and secret are required")
@@ -117,6 +124,7 @@ def connect(  # noqa: PLR0913
         organization_id,
         cloud_provider,
         cloud_region,
+        host,
         database=database or dbname,  # dbname is deprecated.
         statement_results_page_fetch_pause_millis=result_page_fetch_pause_millis,
         http_user_agent=http_user_agent,
@@ -138,7 +146,6 @@ class Connection:
     environment: str
     organization_id: str
     compute_pool_id: str
-    host: str | None
     statement_results_page_fetch_pause_secs: float
     """Maximum seconds to wait between fetching pages of statement
         results (per statement). Prevents tight loops of requests to the
@@ -173,9 +180,9 @@ class Connection:
         environment: str,
         compute_pool_id: str,
         organization_id: str,
-        cloud_provider: str,
-        cloud_region: str,
-        host: str | None = None,
+        cloud_provider: str | None,
+        cloud_region: str | None,
+        host: str | None,
         database: str | None = None,
         statement_results_page_fetch_pause_millis: int = 100,
         http_user_agent: str | None = None,
@@ -191,6 +198,9 @@ class Connection:
             organization_id: Organization ID
             cloud_provider: Cloud provider
             cloud_region: Cloud region (e.g., "us-east-2", "us-west-2")
+            host: The base URL for Confluent Cloud API (optional). If not provided, the host will be
+                constructed based on the cloud_provider and cloud_region parameters in the format
+                "https://flink.{cloud_region}.{cloud_provider}.confluent.cloud".
             result_page_fetch_pause_millis: Milliseconds to possibly wait between fetching pages of
                 statement results. Defaults to 100ms. If most recent fetch of results for a
                 statement was more than this long ago, then no delay will happen when fetching
@@ -204,7 +214,6 @@ class Connection:
         self.environment = environment
         self.compute_pool_id = compute_pool_id
         self.organization_id = organization_id
-        self.host = host
 
         if statement_results_page_fetch_pause_millis < 0:
             raise InterfaceError("result_page_fetch_pause_millis must be non-negative")
@@ -225,9 +234,11 @@ class Connection:
         )
 
         # Create httpx client for making API calls
-        if self.host is None:
-            self.host = f"https://flink.{cloud_region}.{cloud_provider}.confluent.cloud"
-        base_url = f"{self.host}/sql/v1/organizations/{organization_id}/environments/{environment}"
+        if not host:
+            # Construct the host URL based on cloud provider and region.
+            host = f"https://flink.{cloud_region}.{cloud_provider}.confluent.cloud"
+
+        base_url = f"{host}/sql/v1/organizations/{organization_id}/environments/{environment}"
 
         # Create httpx client for making API calls
         basic_auth = httpx.BasicAuth(username=flink_api_key, password=flink_api_secret)